requirements.txt
fastapi==0.136.0
GitPython==3.1.46
google-genai
google-generativeai
hdbscan==0.8.41
llama_parse==0.6.94
nest_asyncio==1.6.0
nltk==3.9.2
numpy==2.4.4
pandas==3.0.2
pgvector==0.4.2
protobuf
pydantic==2.13.2
python-dotenv==1.2.2
python_docx==1.2.0
Requests==2.33.1
scikit_learn==1.8.0
sentence_transformers==5.2.2
spacy==3.8.11
SQLAlchemy==2.0.49
tqdm==4.67.1
umap_learn==0.5.11
uvicorn==0.40.0
psycopg2-binary==2.9.11